+++ R/onattach.R
@@ -5,14 +5,14 @@
       crayon::italic(
         "scPairs: Go Beyond Marker Genes - Discover Synergistic Gene Pairs in scRNA-seq and Spatial Maps."
       ),
-      " R package version", crayon::bold(" 0.1.1."),
+      " R package version", crayon::bold(" 0.1.2."),
       " Available at: https://github.com/zhaoqing-wang/scPairs"
     )
   } else {
     msg <- paste0(
       "Please cite: Wang Z (2026). ",
       "scPairs: Go Beyond Marker Genes - Discover Synergistic Gene Pairs in scRNA-seq and Spatial Maps. ",
-      "R package version 0.1.1. Available at: https://github.com/zhaoqing-wang/scPairs"
+      "R package version 0.1.2. Available at: https://github.com/zhaoqing-wang/scPairs"
     )
   }